# docs/static/openapi.json
{"openapi":"3.1.0","info":{"title":"Pinglow RestAPI","description":"The RestAPI to interact with Pinglow","license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"},"version":"0.8.0"},"paths":{"/check-status/{target_check}":{"get":{"tags":[],"operationId":"get_check_status","parameters":[{"name":"target_check","in":"path","description":"The check for which we would like to know the status","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The last status of the check","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SimpleCheckResultDto"}}}}}}}},"/check/{target_check}/mute":{"delete":{"tags":[],"operationId":"unmute_check","parameters":[{"name":"target_check","in":"path","description":"The check we would like to unmute","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Whether the unmute operation was successful"}}}},"/check/{target_check}/mute?<until>":{"put":{"tags":[],"operationId":"mute_check","parameters":[{"name":"target_check","in":"path","description":"The check we would like to mute","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Whether the mute operation was successful"}}}},"/check/{target_check}/result":{"post":{"tags":[],"operationId":"process_check_result","parameters":[{"name":"target_check","in":"path","description":"The check for which we would like to send a result","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Whether the processing of the check result was successful"}}}},"/checks":{"get":{"tags":[],"operationId":"get_checks","responses":{"200":{"description":"List of checks","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SimpleCheckDto"}}}}}}}},"/performance-data/{target_check}":{"get":{"tags":[],"operationId":"get_performance_data","parameters":[{"name":"target_check","in":"path","description":"The check for which we would like to get the performance data","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The performance data of the check","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"type":"number","format":"float"},"propertyNames":{"type":"string"}},"propertyNames":{"type":"string","format":"date-time"}}}}}}}}}},"components":{"schemas":{"CheckResultStatus":{"type":"string","enum":["Ok","Warning","Critical","CheckError","Pending"]},"ScriptLanguage":{"type":"string","enum":["Python","Bash"]},"SimpleCheckDto":{"type":"object","required":["check_name","passive"],"properties":{"check_name":{"type":"string"},"interval":{"type":["integer","null"],"format":"int64","minimum":0},"passive":{"type":"boolean"}}},"SimpleCheckResultDto":{"type":"object","required":["check_name","passive","output","status"],"properties":{"check_name":{"type":"string"},"notifications_muted":{"type":["boolean","null"]},"notifications_muted_until":{"type":["string","null"],"format":"date-time"},"output":{"type":"string"},"passive":{"type":"boolean"},"status":{"$ref":"#/components/schemas/CheckResultStatus"},"timestamp":{"type":["string","null"],"format":"date-time"}}}},"securitySchemes":{"api_key":{"type":"apiKey","in":"header","name":"x-api-key"}}},"security":[{"api_key":[]}]}